+++ plugwise/constants.py
@@ -123,7 +123,7 @@ P1_LEGACY_MEASUREMENTS: Final[dict[str, UOM]] = {
 # radiator_valve: 'uncorrected_temperature', 'temperature_offset'
 
 DEVICE_MEASUREMENTS: Final[dict[str, DATA | UOM]] = {
-    "humidity": UOM(PERCENTAGE),  # Specific for a Jip
+    "humidity": UOM(NONE),  # Specific for a Jip
     "illuminance": UOM(UNIT_LUMEN),  # Specific for an Anna
     "temperature": UOM(TEMP_CELSIUS),  # HA Core thermostat current_temperature
     "thermostat": DATA("setpoint", TEMP_CELSIUS),  # HA Core thermostat setpoint
@@ -425,44 +425,44 @@ class SmileBinarySensors(TypedDict, total=False):
 class SmileSensors(TypedDict, total=False):
     """Smile Sensors class."""
 
-    battery: float
+    battery: int
     cooling_activation_outdoor_temperature: float
     cooling_deactivation_threshold: float
     dhw_temperature: float
     domestic_hot_water_setpoint: float
     temperature: float
-    electricity_consumed: int
-    electricity_consumed_interval:int
+    electricity_consumed: float
+    electricity_consumed_interval: float
     electricity_consumed_off_peak_cumulative: float
-    electricity_consumed_off_peak_interval: int
-    electricity_consumed_off_peak_point: int
+    electricity_consumed_off_peak_interval: float
+    electricity_consumed_off_peak_point: float
     electricity_consumed_peak_cumulative: float
-    electricity_consumed_peak_interval: int
-    electricity_consumed_peak_point: int
-    electricity_consumed_point: int
-    electricity_phase_one_consumed: int
-    electricity_phase_two_consumed: int
-    electricity_phase_three_consumed: int
-    electricity_phase_one_produced: int
-    electricity_phase_two_produced: int
-    electricity_phase_three_produced: int
-    electricity_produced: int
-    electricity_produced_interval: int
+    electricity_consumed_peak_interval: float
+    electricity_consumed_peak_point: float
+    electricity_consumed_point: float
+    electricity_phase_one_consumed: float
+    electricity_phase_two_consumed: float
+    electricity_phase_three_consumed: float
+    electricity_phase_one_produced: float
+    electricity_phase_two_produced: float
+    electricity_phase_three_produced: float
+    electricity_produced: float
+    electricity_produced_interval: float
     electricity_produced_off_peak_cumulative: float
-    electricity_produced_off_peak_interval: int
-    electricity_produced_off_peak_point: int
+    electricity_produced_off_peak_interval: float
+    electricity_produced_off_peak_point: float
     electricity_produced_peak_cumulative: float
-    electricity_produced_peak_interval: int
-    electricity_produced_peak_point: int
-    electricity_produced_point: int
+    electricity_produced_peak_interval: float
+    electricity_produced_peak_point: float
+    electricity_produced_point: float
     gas_consumed_cumulative: float
     gas_consumed_interval: float
     humidity: float
     illuminance: float
     intended_boiler_temperature: float
-    modulation_level: float
+    modulation_level: int
     net_electricity_cumulative: float
-    net_electricity_point: int
+    net_electricity_point: float
     outdoor_air_temperature: float
     outdoor_temperature: float
     return_temperature: float
@@ -470,7 +470,7 @@ class SmileSensors(TypedDict, total=False):
     setpoint_high: float
     setpoint_low: float
     temperature_difference: float
-    valve_position: float
+    valve_position: int
     voltage_phase_one: float
     voltage_phase_two: float
     voltage_phase_three: float